jquery validation dynamic fields

+++ app/Http/Controllers/ContactController.php
@@ -30,10 +30,10 @@ class ContactController extends Controller
     protected function store(ContactForm $form): JsonResponse
     {
         if ($form->profile_image) {
-            $image = $this->uploadImage();
+            $image = $this->uploadImageGetName();
         }
         $contact = Contact::create(['email' => request('email'), 'profile_image' => $image ?? null]);
-        $contact->addTel($contact, request('telephone_number_new'));
+        $contact->addTel($contact, request('telephone_number'));
         $contact->tags()->sync(request('tags_id'));
         session()->flash('success', 'You have successfully created a contact.');
 
@@ -54,16 +54,18 @@ class ContactController extends Controller
     protected function update(ContactForm $form, Contact $contact): JsonResponse
     {
         if ($form->profile_image) {
-            $image = $this->uploadImage();
-            $image_path = public_path("images/$contact->profile_image");
-            if (file_exists($image_path)) {
-                unlink($image_path);
+            $image = $this->uploadImageGetName();
+            if($contact->profile_image) {
+                $image_path = public_path("images/$contact->profile_image");
+                if (file_exists($image_path)) {
+                    unlink($image_path);
+                }
             }
         }
+
         $contact->update(['email' => request('email'), 'profile_image' => $image ?? $contact->profile_image]);
         $contact->tags()->sync(request('tags_id'));
-        $contact->updateTels(request('telephone_number'));
-        $contact->addTel($contact, request('telephone_number_new'));
+        $contact->updateTels($contact, request('telephone_number'));
         session()->flash('success', 'You have successfully edited a contact.');
 
         return response()->json(['route' => route('contacts.index')]);
@@ -83,7 +85,7 @@ class ContactController extends Controller
         return back()->withSuccess('Successfully deleted contact.');
     }
 
-    private function uploadImage(): string
+    private function uploadImageGetName(): string
     {
         $image = request('profile_image');
         $name = Contact::generateUniqueImageName() . '.' . $image->extension();
@@ -106,6 +108,16 @@ class ContactController extends Controller
         $contact ? $res = false : $res = true;
 
         return response()->json($res);
+    }
+
+    protected function sortTable(Request $request): JsonResponse
+    {
+        $column = $request->get('column');
+        $sort = $request->get('sort');
+        $contacts = Contact::sort($column, $sort);
+        return response()->json([
+             'view' => view('contacts.ajax-table-sort', compact('contacts'))->render()
+        ]);
     }
 
 }
